+++ Commands/Commands/AllPhrases.js
@@ -17,7 +17,8 @@ module.exports = {
   usage: "wfrases <Sin Parámetros>",
   nsfw: false,
   enabled: true,
-  permissions: "",
+  aliases: [],
+  permissions: [],
   async execute(message = new Message(), args = new Array()) {
     const { member, channel, guild } = message;
