+++ Commands/Commands/AnimeFLV.js
@@ -31,7 +31,8 @@ module.exports = {
   usage: "flv [Título del animé]",
   nsfw: false,
   enabled: true,
-  permissions: "",
+  aliases: [],
+  permissions: [],
   async execute(message = new Message(), args = new Array()) {
 
     const { channel, author } = message;
