+++ Commands/Commands/Invite.js
@@ -2,7 +2,7 @@ const { Message, MessageEmbed } = require('discord.js');
 
 const invitation = (invite) =>
   new MessageEmbed()
-    .setTitle(`📩 Invitación:`)
+    .setTitle(`📩 Enlace de Invitación de Muki:`)
     .setDescription(`[Cliquea aquí](${invite})`)
     .setColor("BLUE");
 
@@ -12,7 +12,8 @@ module.exports = {
   usage: "invite <Sin Parámetros>",
   enabled: true,
   nsfw: false,
-
+  aliases: [],
+  permissions: [],
   async execute(message = new Message(), args = new Array()) {
     const { channel, client } = message;
     const invite = await client.generateInvite(607177824);
