+++ Commands/Commands/Kona.js
@@ -47,7 +47,8 @@ module.exports = {
   usage: "kona [tags]",
   nsfw: true,
   enabled: true,
-  permissions: "",
+  aliases: [],
+  permissions: [],
   async execute(message = new Message(), args = new Array()) {
     const { channel, author } = message;
