+++ Commands/Commands/Moan.js
@@ -4,8 +4,8 @@
  */
 const Moaning = {};
 const { Message } = require('discord.js');
-const fs = require('fs');
-const Moans = fs.readdirSync('../Music/Moans').filter(file => file.endsWith('.mp3'));
+//const fs = require('fs');
+//const Moans = fs.readdirSync('../Music/Moans').filter(file => file.endsWith('.mp3'));
 
 module.exports = {
   name: "moan",
@@ -13,7 +13,8 @@ module.exports = {
   usage: "moan <Sin Parámetros>",
   nsfw: false,
   enabled: true,
-  permissions: "SPEAK",
+  aliases: [],
+  permissions: ["SPEAK"],
   async execute(message = new Message(), args = new Array()) {
     const { guild, member } = message;
 
@@ -22,7 +23,7 @@ module.exports = {
     if (!member.voice.channel) return await message.reply("tienes que estar en un canal de voz!");
     if (!member.voice.channel.joinable) return await message.reply("no puedo entrar al canal de voz!\nPor favor revisa mis permisos.");
     const connection = await member.voice.channel.join();
-    const dispatcher = connection.play('Modules/Music/Moans/test.mp3', { volume: 0.50, bitrate: 96000 });
+   // const dispatcher = connection.play('Modules/Music/Moans/test.mp3', { volume: 0.50, bitrate: 96000 });
 
     dispatcher.on('end', () => {
       console.log(`Ended`);
