+++ Commands/Commands/Tags.js
@@ -35,7 +35,8 @@ module.exports = {
   usage: "btags [tag]",
   nsfw: false,
   enabled: true,
-  permissions: "",
+  aliases: [],
+  permissions: [],
 
   async execute(message = new Message(), args = new Array()) {
     const { channel } = message;
