+++ Commands/Commands/WelcomeToggle.js
@@ -26,7 +26,8 @@ module.exports = {
   usage: "wtoggle <Sin Parámetros>",
   nsfw: false,
   enabled: true,
-  permissions: "",
+  aliases: [],
+  permissions: [],
   async execute(message = new Message(), args = new Array()) {
     //In this command, content is irrelevant.
     const { author, guild, member, channel, client } = message;
