--- Commands/NekosLife/Endpoints.js
@@ -1,33 +1,0 @@
-module.exports =
-  ['femdom', 'tickle',
-    'classic', 'ngif',
-    'erofeet', 'meow',
-    'erok', 'poke',
-    'les', 'v3',
-    'hololewd', 'nekoapi_v3.1',
-    'lewdk', 'keta',
-    'feetg', 'nsfw_neko_gif',
-    'eroyuri', 'kiss',
-    '8ball', 'kuni',
-    'tits', 'pussy_jpg',
-    'cum_jpg', 'pussy',
-    'lewdkemo', 'lizard',
-    'slap', 'lewd',
-    'cum', 'cuddle',
-    'spank', 'smallboobs',
-    'goose', 'Random_hentai_gif',
-    'avatar', 'fox_girl',
-    'nsfw_avatar', 'hug',
-    'gecg', 'boobs',
-    'pat', 'feet',
-    'smug', 'kemonomimi',
-    'solog', 'holo',
-    'wallpaper', 'bj',
-    'woof', 'yuri',
-    'trap', 'anal',
-    'baka', 'blowjob',
-    'holoero', 'feed',
-    'neko', 'gasm',
-    'hentai', 'futanari',
-    'ero', 'solo', 'waifu',
-    'pwankg', 'eron', 'erokemo']
